i have footten to execute actions with managedobjectcontext perfromBlock

+++ SimpleTaskManager/Database/Controller/DBController+Internal.h
@@ -18,7 +18,7 @@
 
 - (STMTask *)renameTaskWithId:(NSString *)uid withName:(NSString *)theNewName error:(NSError **)error;
 
-- (STMTask *)reorderTaskWithId:(NSString *)uid toIndex:(NSUInteger)index1 error:(NSError **)error;
+- (STMTask *)reorderTaskWithId:(NSString *)uid toIndex:(NSInteger)index1 error:(NSError **)error;
 
 - (STMTask *)findTaskWithId:(NSString *)uid error:(NSError **)error;
 
@@ -28,5 +28,5 @@
 
 - (NSArray *)fetchAllTasks:(NSError **)error;
 
-- (BOOL)reorderTask:(STMTask *)task withIndex:(int)index error:(NSError **)error;
+- (BOOL)reorderTask:(STMTask *)task withIndex:(NSInteger)index error:(NSError **)error;
 @end
